+++ packages/backend/src/database/entities/savedCharts.ts
@@ -9,6 +9,7 @@ import {
     MetricFilterRule,
     MetricType,
     NumberSeparator,
+    TimeZone,
 } from '@lightdash/common';
 import { Knex } from 'knex';
 
@@ -82,6 +83,7 @@ export type DbSavedChartVersion = {
     chart_config: ChartConfig['config'] | undefined;
     pivot_dimensions: string[] | undefined;
     updated_by_user_uuid: string | undefined;
+    timezone: TimeZone | undefined;
 };
 
 export type SavedChartVersionsTable = Knex.CompositeTableType<
@@ -99,6 +101,7 @@ export type CreateDbSavedChartVersion = Pick<
     | 'pivot_dimensions'
     | 'chart_config'
     | 'updated_by_user_uuid'
+    | 'timezone'
 >;
 
 type DbSavedChartVersionField = {