+++ packages/backend/src/models/SavedChartModel.ts
@@ -20,6 +20,7 @@ import {
     SessionUser,
     SortField,
     Space,
+    TimeZone,
     UpdatedByUser,
     UpdateMultipleSavedChart,
     UpdateSavedChart,
@@ -71,6 +72,7 @@ type DbSavedChartDetails = {
     last_name: string;
     pinned_list_uuid: string;
     dashboard_uuid: string | null;
+    timezone: TimeZone | undefined;
 };
 
 const createSavedChartVersionField = async (
@@ -137,6 +139,7 @@ const createSavedChartVersion = async (
             tableCalculations,
             additionalMetrics,
             customDimensions,
+            timezone,
         },
         chartConfig,
         tableConfig,
@@ -155,6 +158,7 @@ const createSavedChartVersion = async (
                 chart_type: chartConfig.type,
                 chart_config: chartConfig.config,
                 updated_by_user_uuid: updatedByUser?.userUuid,
+                timezone,
             })
             .returning('*');
         const promises: Promise<any>[] = [];
@@ -632,6 +636,7 @@ export class SavedChartModel {
                     'saved_queries_versions.created_at',
                     'saved_queries_versions.chart_config',
                     'saved_queries_versions.pivot_dimensions',
+                    'saved_queries_versions.timezone',
                     `${OrganizationTableName}.organization_uuid`,
                     `${OrganizationTableName}.chart_colors`,
                     `${UserTableName}.user_uuid`,
@@ -827,6 +832,7 @@ export class SavedChartModel {
                         binWidth: cd.bin_width,
                         customRange: cd.custom_range,
                     })),
+                    timezone: savedQuery.timezone,
                 },
                 chartConfig,
                 tableConfig: {