+++ packages/common/src/types/featureFlags.ts
@@ -45,4 +45,13 @@ export enum FeatureFlags {
 
     /** Enable dashboard tabs */
     DashboardTabs = 'dashboard_tabs',
+
+    /**
+     * Disable https://docs.snowflake.com/en/sql-reference/parameters#label-quoted-identifiers-ignore-case
+     * for the Snowflake warehouse client
+     */
+    DisableSnowflakeQuotedIdentifiersIgnoreCase = 'disable-snowflake-quoted-identifiers-ignore-case',
+
+    /* Send local timezone to the warehouse session */
+    EnableUserTimezones = 'enable-user-timezones',
 }