+++ packages/warehouses/src/warehouseClients/TrinoWarehouseClient.ts
@@ -169,13 +169,21 @@ export class TrinoWarehouseClient extends WarehouseBaseClient<CreateTrinoCredent
         };
     }
 
-    async runQuery(sql: string, tags?: Record<string, string>) {
+    async runQuery(
+        sql: string,
+        tags?: Record<string, string>,
+        timezone?: string,
+    ) {
         const { session, close } = await this.getSession();
         let query: Iterator<QueryResult>;
         try {
             let alteredQuery = sql;
             if (tags) {
                 alteredQuery = `${alteredQuery}\n-- ${JSON.stringify(tags)}`;
+            }
+            if (timezone) {
+                console.debug(`Setting Trino timezone to ${timezone}`);
+                await session.query(`SET TIME ZONE '${timezone}'`);
             }
             query = await session.query(sql);
 
@@ -231,7 +239,7 @@ export class TrinoWarehouseClient extends WarehouseBaseClient<CreateTrinoCredent
                 } catch (e: any) {
                     throw new WarehouseQueryError(e.message);
                 } finally {
-                    if (query) close();
+                    if (query) void close();
                 }
             });
 
@@ -240,10 +248,6 @@ export class TrinoWarehouseClient extends WarehouseBaseClient<CreateTrinoCredent
             await close();
         }
         return catalogToSchema(results);
-    }
-
-    getFieldQuoteChar() {
-        return '"';
     }
 
     getStringQuoteChar() {